made work for mac

+++ base/StringUtil.cc
@@ -20,7 +20,7 @@ int IsSpace(char c) {
 }
 
 int ToWords(string &orig, vector<string> &words) {
-    int curWordStart, curWordEnd;
+    size_t curWordStart, curWordEnd;
     curWordStart = 0;
     while(curWordStart < orig.size()) {
         while (curWordStart < orig.size() and IsSpace(orig[curWordStart])) {
